requirements.txt
# Reachy MVP Requirements

# Core dependencies
numpy>=1.24.0
pyaudio>=0.2.13
pyyaml>=6.0

# OpenAI
openai>=1.10.0
python-dotenv>=1.0.0

# Voice Activity Detection
onnxruntime>=1.16.0

# Speech-to-Text
faster-whisper>=1.0.0

# Wake word detection
openwakeword>=0.6.0

# Reachy robot
reachy-mini>=1.0.0

# HTTP requests (for TTS service)
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
